% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topictile_mode.R
\name{get_topictile_mode_from_model}
\alias{get_topictile_mode_from_model}
\title{Creates a topictile mode given a topic model}
\usage{
get_topictile_mode_from_model(model, gibbs_iters, dst)
}
\arguments{
\item{model}{A topic model.}

\item{gibbs_iters}{An integer specifying how many sampling to do.}

\item{dst}{A filename to where the model will write its samples temporarily.}
}
\description{
Creates a topictile mode given a topic model that generates \code{gibbs_iters} samples.
}
