% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topictile_mode.R
\name{read_topictile_mode_from_file}
\alias{read_topictile_mode_from_file}
\title{Reads a topictile mode from file}
\usage{
read_topictile_mode_from_file(src)
}
\arguments{
\item{src}{A filename to where the mode data will be read.}
}
\description{
Reads a topictile mode from specified file location.
}
