% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{token_seg}
\alias{token_seg}
\title{Token Segmentizer}
\usage{
token_seg(corpus, seg_size, tokenized_corpus = NULL, token = "words", ...)
}
\description{
Segments a corpus based on a fixed number of tokens
}
