% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topictile_mode.R
\name{write_topictile_mode_to_file}
\alias{write_topictile_mode_to_file}
\title{Writes a topictile mode to file}
\usage{
write_topictile_mode_to_file(mode, dst)
}
\arguments{
\item{mode}{A topictile mode.}

\item{dst}{A filename to where the mode data will be written.}
}
\description{
Writes a topictile mode to specified file location.
}
