% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topictile.R
\name{tf_topictile}
\alias{tf_topictile}
\title{Segmenting text based on Latent Dirichlet Allocation.}
\usage{
tf_topictile(corpus, stopwords, model, gibbs_iters = 50, sentence_size = 20,
  block_size = 2, n_segments = 0, liberal = TRUE)
}
\description{
Segmenting text based on Latent Dirichlet Allocation.
}
